refactoring filter

+++ buisness/market/src/main/kotlin/net/sergey/kosov/market/domains/view/wrappers/ProductFilter.kt
@@ -1,6 +1,7 @@
 package net.sergey.kosov.market.domains.view.wrappers
 
 import net.sergey.kosov.common.annotations.NoArgs
+import java.math.BigDecimal
 
 @NoArgs
-data class ProductFilter(val priceRight: Int, val priceLeft: Int, val title: String)
+data class ProductFilter(val priceRight: BigDecimal, val priceLeft: BigDecimal, val title: String)
